reorder filter and simplify remove label
