+++ storage/postgres/labels.go
@@ -20,6 +20,7 @@ import (
 	"context"
 	"database/sql"
 	"fmt"
+	"strings"
 
 	"github.com/Peripli/service-manager/pkg/query"
 
@@ -66,28 +67,92 @@ func addLabel(ctx context.Context, f func(labelID string, labelKey string, label
 
 func removeLabel(ctx context.Context, execer sqlx.ExecerContext, labelsTableName, referenceColumnName, referenceID, labelKey string, labelValues ...string) error {
 	baseQuery := fmt.Sprintf("DELETE FROM %s WHERE key=$1 AND %s=$2", labelsTableName, referenceColumnName)
-	if len(labelValues) == 0 {
-		return removeAllLabelsWithKey(ctx, execer, baseQuery, referenceID, labelKey)
+	labelValuesCount := len(labelValues)
+	// remove all labels with this key
+	if labelValuesCount == 0 {
+		return execute(ctx, baseQuery, func() (sql.Result, error) {
+			return execer.ExecContext(ctx, baseQuery, labelKey, referenceID)
+		})
 	}
-	baseQuery += " AND val=$3"
-	return removeLabelValues(ctx, execer, baseQuery, referenceID, labelKey, labelValues)
+	// remove labels with a specific key and a value which is in the provided list
+	baseQuery += " AND val IN $3"
+	sqlQuery, queryParams, err := sqlx.In(baseQuery, referenceID, labelKey, labelValues)
+	if err != nil {
+		return err
+	}
+	return execute(ctx, sqlQuery, func() (sql.Result, error) {
+		return execer.ExecContext(ctx, sqlQuery, queryParams)
+	})
 }
 
-func removeLabelValues(ctx context.Context, execerContext sqlx.ExecerContext, query, referenceID, labelKey string, labelValues []string) error {
-	for _, labelValue := range labelValues {
-		if err := execute(ctx, query, func() (sql.Result, error) {
-			return execerContext.ExecContext(ctx, query, labelKey, referenceID, labelValue)
-		}); err != nil {
-			return err
+func buildListQueryWithParams(sqlQuery string, baseTableName string, labelsTableName string, criteria []query.Criterion) (string, []interface{}, error) {
+	if len(criteria) == 0 {
+		return sqlQuery, nil, nil
+	}
+
+	var queryParams []interface{}
+	var queries []string
+
+	sqlQuery += " WHERE "
+	for _, option := range criteria {
+		rightOpBindVar, rightOpQueryValue := buildRightOp(option)
+		sqlOperation := translateOperationToSQLEquivalent(option.Operator)
+		if option.Type == query.LabelQuery {
+			queries = append(queries, fmt.Sprintf("%[1]s.key = ? AND %[1]s.val %[2]s %s", labelsTableName, sqlOperation, rightOpBindVar))
+			queryParams = append(queryParams, option.LeftOp)
+		} else {
+			clause := fmt.Sprintf("%s.%s %s %s", baseTableName, option.LeftOp, sqlOperation, rightOpBindVar)
+			if option.Operator.IsNullable() {
+				clause = fmt.Sprintf("(%s OR %s.%s IS NULL)", clause, baseTableName, option.LeftOp)
+			}
+			queries = append(queries, clause)
+		}
+		queryParams = append(queryParams, rightOpQueryValue)
+	}
+	sqlQuery += strings.Join(queries, " AND ") + ";"
+
+	if hasMultiVariateOp(criteria) {
+		var err error
+		// sqlx.In requires question marks(?) instead of positional arguments (the ones pgsql uses) in order to map the list argument to the IN operation
+		if sqlQuery, queryParams, err = sqlx.In(sqlQuery, queryParams...); err != nil {
+			return "", nil, err
 		}
 	}
-	return nil
+	return sqlQuery, queryParams, nil
 }
 
-func removeAllLabelsWithKey(ctx context.Context, execerContext sqlx.ExecerContext, query, referenceID, key string) error {
-	return execute(ctx, query, func() (sql.Result, error) {
-		return execerContext.ExecContext(ctx, query, key, referenceID)
-	})
+func buildRightOp(criterion query.Criterion) (string, interface{}) {
+	rightOpBindVar := "?"
+	var rhs interface{} = criterion.RightOp[0]
+	if criterion.Operator.IsMultiVariate() {
+		rightOpBindVar = "(?)"
+		rhs = criterion.RightOp
+	}
+	return rightOpBindVar, rhs
+}
+
+func hasMultiVariateOp(criteria []query.Criterion) bool {
+	for _, opt := range criteria {
+		if opt.Operator.IsMultiVariate() {
+			return true
+		}
+	}
+	return false
+}
+
+func translateOperationToSQLEquivalent(operator query.Operator) string {
+	switch operator {
+	case query.LessThanOperator:
+		return "<"
+	case query.GreaterThanOperator:
+		return ">"
+	case query.NotInOperator:
+		return "NOT IN"
+	case query.EqualsOrNilOperator:
+		return "="
+	default:
+		return strings.ToUpper(string(operator))
+	}
 }
 
 func execute(ctx context.Context, query string, f func() (sql.Result, error)) error {
